refactor: isNotInteger 메서드의 위치 변경
Validator에서 유일하게 사용되고 있으므로 가까운 위치로 옮겼다.
